refactor: split onboard success render
